refactor(build): rename 'werf meta-repo disable' to 'meta-repo detach'
The command only deletes the per-project marker from --repo; it neither
moves metadata back nor turns any feature off. "disable" invited both
readings — a user expecting a reverse migration would remove the
safeguard and then let a run without --meta-repo read stale metadata,
which is exactly the failure the safeguard exists to prevent. "detach"
names the actual effect.

Rename the package directory and file, the cobra Use, the root
registration and its import alias, the two error strings in
pkg/storage/meta_repo_marker.go that instruct the user to run the
command, and the en/ru prose in usage/build/process.md. Behaviour is
unchanged and no alias is registered: the command is new in this
unreleased branch, so nothing depends on the old spelling. The
generated CLI reference is regenerated separately.
